split sides in multiples files
